+++ TotalImage.IO/FileSystems/ISO/IsoPrimaryVolumeDescriptor.cs
@@ -178,7 +178,7 @@ namespace TotalImage.FileSystems.ISO
         /// <param name="identifier">The volume descriptor identifier</param>
         /// <param name="version">The version of the volume descriptor</param>
         public IsoPrimaryVolumeDescriptor(in ReadOnlySpan<byte> record, in IsoVolumeDescriptorType type, in ImmutableArray<byte> identifier, in byte version)
-            : base(type, identifier, version)
+            : base((byte)type, identifier, version)
         {
             bool isUnicode = false;
             if (type == IsoVolumeDescriptorType.SupplementaryVolumeDescriptor)
